LPS-96730 Remove portlet preferences for portlets which were removed from the experience

+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/portlet/action/DeleteSegmentsExperienceMVCActionCommand.java
@@ -14,7 +14,10 @@
 
 package com.liferay.layout.content.page.editor.web.internal.portlet.action;
 
+import com.liferay.fragment.model.FragmentEntryLink;
+import com.liferay.fragment.processor.PortletRegistry;
 import com.liferay.fragment.service.FragmentEntryLinkLocalService;
+import com.liferay.fragment.service.FragmentEntryLinkService;
 import com.liferay.layout.content.page.editor.constants.ContentPageEditorPortletKeys;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
@@ -26,16 +29,20 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import com.liferay.portal.kernel.service.PortletPreferencesLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.transaction.Propagation;
 import com.liferay.portal.kernel.transaction.TransactionConfig;
 import com.liferay.portal.kernel.transaction.TransactionInvokerUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.segments.constants.SegmentsConstants;
 import com.liferay.segments.service.SegmentsExperienceService;
+import com.liferay.segments.util.SegmentsExperiencePortletUtil;
 
+import java.util.List;
 import java.util.concurrent.Callable;
 
 import javax.portlet.ActionRequest;
@@ -65,8 +72,12 @@ public class DeleteSegmentsExperienceMVCActionCommand
 			actionRequest, "segmentsExperienceId",
 			SegmentsConstants.SEGMENTS_EXPERIENCE_ID_DEFAULT);
 
-		if (segmentsExperienceId !=
-				SegmentsConstants.SEGMENTS_EXPERIENCE_ID_DEFAULT) {
+		boolean deleteSegmentsExperience = ParamUtil.getBoolean(
+			actionRequest, "deleteSegmentsExperience");
+
+		if (deleteSegmentsExperience &&
+			(segmentsExperienceId !=
+				SegmentsConstants.SEGMENTS_EXPERIENCE_ID_DEFAULT)) {
 
 			_segmentsExperienceService.deleteSegmentsExperience(
 				segmentsExperienceId);
@@ -79,8 +90,42 @@ public class DeleteSegmentsExperienceMVCActionCommand
 			long[] toFragmentEntryLinkIds = JSONUtil.toLongArray(
 				JSONFactoryUtil.createJSONArray(fragmentEntryLinkIdsString));
 
-			_fragmentEntryLinkLocalService.deleteFragmentEntryLinks(
-				toFragmentEntryLinkIds);
+			for (long fragmentEntryLinkId : toFragmentEntryLinkIds) {
+				FragmentEntryLink fragmentEntryLink =
+					_fragmentEntryLinkLocalService.getFragmentEntryLink(
+						fragmentEntryLinkId);
+
+				List<String> portletIds =
+					_portletRegistry.getFragmentEntryLinkPortletIds(
+						fragmentEntryLink);
+
+				for (String portletId : portletIds) {
+					String portletIdWithExperience =
+						SegmentsExperiencePortletUtil.setSegmentsExperienceId(
+							portletId, segmentsExperienceId);
+
+					_portletPreferencesLocalService.deletePortletPreferences(
+						PortletKeys.PREFS_OWNER_ID_DEFAULT,
+						PortletKeys.PREFS_OWNER_TYPE_LAYOUT,
+						fragmentEntryLink.getClassPK(),
+						portletIdWithExperience);
+
+					if (segmentsExperienceId ==
+							SegmentsConstants.SEGMENTS_EXPERIENCE_ID_DEFAULT) {
+
+						_portletPreferencesLocalService.
+							deletePortletPreferences(
+								PortletKeys.PREFS_OWNER_ID_DEFAULT,
+								PortletKeys.PREFS_OWNER_TYPE_LAYOUT,
+								fragmentEntryLink.getClassPK(), portletId);
+					}
+				}
+			}
+
+			if (deleteSegmentsExperience) {
+				_fragmentEntryLinkLocalService.deleteFragmentEntryLinks(
+					toFragmentEntryLinkIds);
+			}
 		}
 	}
 
@@ -124,6 +169,15 @@ public class DeleteSegmentsExperienceMVCActionCommand
 
 	@Reference
 	private FragmentEntryLinkLocalService _fragmentEntryLinkLocalService;
+
+	@Reference
+	private FragmentEntryLinkService _fragmentEntryLinkService;
+
+	@Reference
+	private PortletPreferencesLocalService _portletPreferencesLocalService;
+
+	@Reference
+	private PortletRegistry _portletRegistry;
 
 	@Reference
 	private SegmentsExperienceService _segmentsExperienceService;
